pyproject.toml
[tool.poetry]
name = "oceanstart"
version = "0.1.0"
description = "Test app"
authors = ["Leva <smj510black@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
neovim = "^0.3.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
